Throw a compiler error for unsupported comparison types (#4)
Most comparison types are assumed to be compatible with any type of
literal since it is up to the specification generator to interpret
the comparison.

However some types of comparison can never make logical
sense; specifically relative comparisons such as greater-than and
less-than which do not make sense with boolean operands. It does not
make sense to say that something is greater than true or less than
false for example. Therefore this commit causes a CompilerError to
be thrown when boolean literals are compared using any operator other
than equals.

In future this can be extended if any other illogical comparisons
should be prevented.

Additionally, this commit upgrades PHPUnit to version 8.2 and makes
the necessary fixes to tests.