Move restricted boolean operators from compiler and into specifications
This is consistent with other comparison operator and value combinations
in that the specification (or specification generator) decides what is
allowed and what is not. Even though it is unlikely anyone would want to
support operators other than EQUALS with boolean values, the option is
provided for other combinations and it's better to be consistent.