Compile `in` comparisons to a single logical OR specification
A composite specification of type logical OR is generated. This is
satisfied when any of its children is satisfied. The compiler generates
children by converting the `in` expression to a set of `is` specifications.

For example, given the expression `foo in [1,2,3]`, 3 child
specifications are produced as if the expression
`foo is 1 or foo is 2 or foo is 3` was compiled.

This removes some complexity from specifications themselves since they
no longer need to handle the `in` operator. However this behaviour can
be disabled if it is preferable that specifications are generated as
normal for `in` comparisons. This might be the case if a checking
many individual specifications would be more expensive for example.